aucMRMCcluster {mitoticFigureCounts}	R Documentation

aucMRMCcluster

Description

The results of a multi-reader multi-case (MRMC) analysis of the auc for each scanner. The MRMC analysis is accomplished by the OR method (Obuchowski and Rockette: Obuchowski1995_Commun-Stat-Simulat_v24p285). The function used is mrmcAnalysisOR. Since the data is binary, auc is the average of sensitivity and specificity or half of (Youden's index + 1). Sensitivity is defined as the number of MFs detected by an observer divided by the number of true MFs. Specificity is defined as one minus the false-positive fraction, where the false-positive fraction is the number of false MFs that were positively marked, divided by the total number of false MFs. Furthermore, we account for the fact that there are multiple observations per case (multiple ROIs per WSI, clustered data: Obuchowski1997_Biometrics_v53p567) when calculating the reader by modality covariances that are used in the OR method. The function used is doAUCcluster

Usage


aucMRMCcluster

Format

An object of class list of length 4.

Details

The results of this script are a list of MRMC AUC analysis results for 4 scanners.

Scanner.A,

Scanner.B,

Scanner.C,

Scanner.D

Each scanner results is itself a list contains the MRMC analysis results for the null hypothesis that the scanner accuracy is the same as the microscope accuracy:

F_stat, F-statistic

df.H, Degrees of freedom

p, p-value

theta.i, [2] Accuracy of the microscope and scanner

df.sgl, [2] Degrees of freedom for the microscope and the scanner

se.i, [2] Standard error of the microscope and scanner accuracy

se.dif, Standard erro of the microscope and scanner accuracy difference

covOR, [12] Obuchowski and Rockette covariances: varR, varTR, cov1, cov2, cov3, varE, varR.1, varR.2, cov2.1, cov2.2, varE.1, varE.2

theta.hat, [2x5] Accuracy of each reader for the microscope (row 1) and accuracy of each reader for the scanner (row 2)

cov.hat, [10x10] Reader by modality (microscope and scanner) covariance matrix

botCI, [2] Bottom 95

topCI, [2] Top 95

#' The analysis results are used to produce:

Table 4: Accuracy for all readers and observation methods

Figure 3: Accuracy (average of sensitivity and specificity) for each viewing mode averaged over all the readers with 95 The asterisks indicate that the difference in accuracy of the viewing mode compared to that of microscopy is statistically significant. All analyses account for the correlations and variability from the readers reading the same ROIs.

[Package mitoticFigureCounts version 1.0 Index]