dfClassify20180627 {mitoticFigureCounts}	R Documentation

dfClassify20180627

Description

A single data frame of the study data. Each row corresponds to a candidate mitotic figure and modality (155 candidates * 5 modalities = 775 rows)

Usage


dfClassify20180627

Format

An object of class data.frame with 775 rows and 14 columns.

Details

A data frame with 14 columns:

targetID [Factor] Target cell ID = (candidate mitotic figure ID)

cell.mark [Factor] ROI ID, "dash", Cell ID in the ROI

wsiName [Factor] WSI file name

roiID [Factor] ROI ID

cellID.mskcc20171103 [Factor] Descriptive target cell ID, including ROI ID, WSI file name, x position of target cell, y position of target cell

xCell [Factor] x position of target cell

yCell [Factor] y position of target cell

observer.1 [num] Observer1 modality-specific decision on the candidate MF

observer.2 [num] Observer2 modality-specific decision on the candidate MF

observer.3 [num] Observer3 modality-specific decision on the candidate MF

observer.4 [num] Observer4 modality-specific decision on the candidate MF

observer.5 [num] Observer5 modality-specific decision on the candidate MF

truth [num] Truth panel decision on the candidate MF

modalityID [Factor] modality ID (scanner or microscope)

Some of this data is missing location information.

[Package mitoticFigureCounts version 1.0 Index]