dfCountROI20180627 {mitoticFigureCounts}	R Documentation

dfCountROI20180627

Description

A single data frame of the mitotic figure counts per ROI and modality (40 ROIs x 5 modalities = 200 rows). There is a column for each observer and the truth.

Usage


dfCountROI20180627

Format

An object of class data.frame with 200 rows and 9 columns.

Details

A data frame with 9 columns:

wsiName [Factor] WSI file name

roiID [Factor] ROI ID

modalityID [Factor] modality, scanner ID

observer.1 [num] Observer1 modality-specific count for the ROI

observer.2 [num] Observer2 modality-specific count for the ROI

observer.3 [num] Observer3 modality-specific count for the ROI

observer.4 [num] Observer4 modality-specific count for the ROI

observer.5 [num] Observer5 modality-specific count for the ROI

truth [num] Truth panel count for the ROI

[Package mitoticFigureCounts version 1.0 Index]