dfCountWSI20180627 {mitoticFigureCounts}	R Documentation

dfCountWSI20180627

Description

A single data frame of the mitotic figure counts per WSI and modality (4 WSIs x 5 modalities = 20 rows). There is a column for each observer and the truth.

Usage


dfCountWSI20180627

Format

An object of class data.frame with 20 rows and 8 columns.

Details

A data frame with 8 columns:

wsiName [Factor] WSI file name

modalityID [Factor] modality, scanner ID

observer.1 [num] Observer1 modality-specific count for the WSI

observer.2 [num] Observer2 modality-specific count for the WSI

observer.3 [num] Observer3 modality-specific count for the WSI

observer.4 [num] Observer4 modality-specific count for the WSI

observer.5 [num] Observer5 modality-specific count for the WSI

truth [num] Truth panel count for the WSI

[Package mitoticFigureCounts version 1.0 Index]