doAUCcluster {mitoticFigureCounts}	R Documentation

doAUCcluster

Description

Do AUC analysis of clustered data

Usage


doAUCcluster(predictor1, predictor2 = NULL, response, clusterID,
  alpha = 0.05, level = NULL, print.all = F)

Arguments

predictor1
a vector containing the predictor for ROC curve 1

predictor2
a vector containing the predictor for ROC curve 2

response
a vector containing the response for both ROC curves

clusterID
a vector containing IDs for the clusters

alpha
the type I error rate

level
can be used to specify the response level considered positive (if omitted, the second level of the response is selected)

print.all
if TRUE, intermediate estimates are printed

Details

This function is used to perform the statistical methods in: Obuchowski NA. "Nonparametric analysis of clustered ROC curve data." Biometrics. 1997: 567-578.

This file was downloaded from the Cleveland Clinic Lerner Research Institute Department of Quantitative Health Sciences Software web page.

FILE: https://www.lerner.ccf.org/qhs/software/lib/funcs_clusteredROC.R

WEBPAGE: https://www.lerner.ccf.org/qhs/software/roc_analysis.php

iMRMC users shared the links during a discussion with questions about how to analyze MRMC data that was clustered. https://github.com/DIDSR/iMRMC/issues/147 There is a short pdf tutorial a https://www.lerner.ccf.org/qhs/software/lib/clusteredROC_help.pdf. It exists in the inst/extra/docs folder of the repository. It exists in the extra/docs folder of the installed package.

Value

[list] auc, auc.se, ci.for.auc

[Package mitoticFigureCounts version 1.0 Index]