mrmcAnalysisOR {mitoticFigureCounts}	R Documentation

MRMC analysis by Obuchowski and Rockette ( Obuchowski1995_Commun-Stat-Simulat_v24p285)

Description

OR's method to estimate the MRMC variance for a given theta. This function requires an nModalities by nReaders covariance matrix to estimate covOR = (cov1, cov2, cov3), as described by Hillis (2014, SIM, Section 2.2)

Usage


mrmcAnalysisOR(theta.hat, cov.hat, sgl = "corres")

Arguments

theta.hat
[2, nReaders]: Performance estimates for two modalities by all readers

cov.hat
[2 * nReaders, 2 * nReaders]: Covariance matrix between each reader x modality

sgl
ch: Flat indicating whether or not to do single modality analysis only using single modality data

Value

list $ F : F statistic $ df.H : degrees of freedom difference of concordance $ p : P value $ se.dif : difference of variance $ theta.i [2] : reader-averaged HarrellsC form modality A & B $ df.sgl [2] : degrees of freedom for modality A & B $ se.i [2] : Variance form modality A & B $ covOR [6] : Components of variance of the Obuchowski and Rockette method cov1, cov2(pooled over modalities), cov3, varC, cov2(modalityA only), cov2(modalityB only)

[Package mitoticFigureCounts version 1.0 Index]